Chemostat solver, internal constituents and bug fixes
Implementing extensive chemostat solver that allows for variable reactor
volume and includes agent mass and internal constituents in ODE. Note:
if agent constituents are more complex than simply the mass of the
agents these constituents have to be saved in a massMap aspect. The
massMap can be used to calculate the agent mass (as a sum of it's parts)
if required drymass can be converted to wetmass.