Commit includes accumulated fixes and updates:
- Added (UpdateBodySurfaceScaled) agent scaling for 2D representation of
3D shapes to maintain at constant volume. Coccoid cells are calculated
using a cylinder shape aligning the domain z-axis. rods are calculated
using one cylinder + a box were the length is set so the total volume
equals the volume of the represented capsule.
- Streamlined total agent mass calculation in multiple classes (Method
placed in Helper)
- Introducing agent differentiation, allowing for agents to switch
species definition when a certain user settable boolean expression is
met. This includes a minor adjustment of the FilterLogic class which is
used to evaluate the users boolean expression and the possibility to
replace an agents species module with an alternative one.
- Some additions to the (testing) gravity implementation within the
AgentRelaxation class, now a BiofilmCondensationLimit can be set to
allow for a condensing force during a limited amount of time steps (user
defined) in simulations where starvation leads to huge void regions (NB:
Lets discuss implications and implementation next meeting).
- Some bug fixes (including in InstantiableMap, check InstantiableList
is todo).
- Test protocol files that implement mentioned features.