Added option to disable storing data to disc + small bugfix.
Running Unit tests creates a lot of files on disc that are typically not
useful as we are simply trying to test for past/no-pass conditions. This
update allows us to disable writing to disc for specific simulation
avoiding having to manually clean-up after. Further I encountered a
small bug in the handling of mass maps in CoccoidDivision.