Patched an issue violating mass conservation
Consumption trough a biofilm boundary was resolved in the next time-step
essentially causing a delay between compartments and resulting in
consumption of the last time-step to be ignored completely. This results
in a significant violation of the conservation principle for shorter
simulations.

The delay is now mediated by first obtaining biofilm consumption and
then updating the chemostat boundary resulting in appropriate and
simultaneous consumption in the chemostat compartment. To ensure correct
behavior the biofilm compartment must have a priority integer which is
lower than that of the chemostat compartment.